simplified the ssr server files and webpack config, somewhat.
