+++ api/package.json
@@ -110,9 +110,9 @@
     "chai": "^3.5.0",
     "chai-as-promised": "^5.3.0",
     "chance": "^1.0.4",
-    "eslint": "^3.2.0",
+    "eslint": "^3.2.2",
     "eslint-config-boldr": "^1.0.2",
-    "mocha": "^2.5.3",
+    "mocha": "^3.0.0",
     "nodemon": "^1.10.0",
     "sequelize-cli": "^2.4.0",
     "sinon": "^1.17.5",
