+++ cms/package.json
@@ -193,9 +193,11 @@
     "babel-core": "^6.11.4",
     "babel-eslint": "6.1.2",
     "babel-loader": "^6.2.4",
+    "babel-plugin-lodash": "^3.2.6",
     "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.12.0",
+    "babel-plugin-webpack-loaders": "^0.7.1",
     "babel-polyfill": "^6.9.1",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-es2015-native-modules": "^6.9.2",
@@ -211,7 +213,7 @@
     "cssnano": "^3.7.3",
     "dotenv": "2.0.0",
     "enzyme": "^2.4.1",
-    "eslint": "^3.2.0",
+    "eslint": "^3.2.2",
     "eslint-config-boldr": "^1.0.2",
     "expect": "^1.20.2",
     "extract-text-webpack-plugin": "^2.0.0-beta.3",
@@ -223,6 +225,7 @@
     "jsdom": "^9.4.1",
     "jsdomify": "^2.1.0",
     "json-loader": "0.5.4",
+    "lodash-webpack-plugin": "^0.9.3",
     "mocha": "^3.0.0",
     "nock": "^8.0.0",
     "node-sass": "^3.8.0",
@@ -252,7 +255,8 @@
     "url-loader": "^0.5.7",
     "webpack": "^2.1.0-beta.20",
     "webpack-dev-middleware": "1.6.1",
-    "webpack-hot-middleware": "2.12.2"
+    "webpack-hot-middleware": "2.12.2",
+    "why-did-you-update": "0.0.8"
   },
   "stylelint": {
     "rules": {
