+++ cms/src/scenes/Boldr/Boldr.jsx
@@ -1,0 +1,2 @@
+import React from 'react';
+import { connect } from 'react-redux';
