--- cms/src/server/core/index.js
@@ -1,7 +1,0 @@
-import boldrSSR from './middleware/boldrSSR';
-import webserver from './webserver';
-
-export {
-  boldrSSR,
-  webserver
-};
