+++ cms/src/server/middleware.js
@@ -1,7 +1,6 @@
 import path from 'path';
 import express from 'express';
 import bodyParser from 'body-parser';
-import methodOverride from 'method-override';
 import cors from 'cors';
 import compression from 'compression';
 
@@ -11,12 +10,7 @@ export default app => {
   app.use(cors({ origin: true, credentials: true }));
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: true }));
-  app.use(methodOverride('X-HTTP-Method-Override'));
   app.use(express.static(path.resolve('public')));
   app.set('trust proxy', 'loopback');
   app.options('*', (req, res) => res.sendStatus(200));
-
-  console.log('--------------------------');
-  console.log('===> 😊  Starting Boldr SSR. . .');
-  console.log(`===> 🌎  Listening on ${process.env.SSR_PORT} with environment: ${process.env.NODE_ENV}`);
 };
