+++ cms/src/server/server.js
@@ -1,10 +1,9 @@
 import Debug from 'debug';
 import express from 'express';
 import errorHandler from 'errorhandler';
-import Boom from 'boom';
 import httpProxy from 'http-proxy';
-
-import { webserver, boldrSSR } from './core';
+import middleware from './middleware';
+import boldrSSR from './boldrSSR';
 
 const debug = Debug('boldrCMS:engine');
 const targetUrl = process.env.TARGET_URL;
@@ -16,8 +15,8 @@ const proxy = httpProxy.createProxyServer({
 });
 
 debug('express middleware');
-webserver(app);
-// 
+middleware(app);
+
 // app.use('/api/v1', (req, res) => {
 //   const url = targetUrl + req.originalUrl;
 //   debug(`${req.method} PROXY to ${url}`);
@@ -34,15 +33,9 @@ app.use(boldrSSR);
 
 // Use Boom for 404 error handling.
 app.use((req, res, next) => {
-  next(Boom.notFound('Looks like you might be lost...'));
+  next();
 });
 
-// Wrap other errors with Boom.
-app.use((err, req, res, next) => {
-  const { statusCode, payload } = Boom.wrap(err).output;
-  res.status(statusCode).json(payload);
-  next(err);
-});
 // Handle our errors.
 app.use(errorHandler);
