+++ cms/tools/webpack/util/createSourceLoader.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+export default function createSourceLoader(spec) {
+  return Object.keys(spec).reduce((x, key) => {
+    x[key] = spec[key];
+
+    return x;
+  }, {
+    include: [path.resolve(process.cwd(), 'src')]
+  });
+}
